requirements.txt
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
idna==2.10
numpy==1.19.4
pandas==1.1.4
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.20
tqdm==4.53.0
urllib3==1.26.2
APScheduler~=3.6.3
aiogram==2.9.2
aiohttp==3.6.2
aioredis==1.3.1
python-dotenv
